requirement.txt
Jinja2==2.7.1
#PIL==1.1.7
pillow
gunicorn==18.0
pymongo==2.6.3
redis==2.9.0
requests==2.1.0
sinaweibopy==1.1.3
web.py==0.37
lxml